% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_braces_to_ontology_terms}
\alias{add_braces_to_ontology_terms}
\title{A little helper function in case ontology terms are not udpated}
\usage{
add_braces_to_ontology_terms(x)
}
\arguments{
\item{x}{vector of ontology terms to update}
}
\description{
A little helper function in case ontology terms are not udpated
}
