% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{all_table_ontology_columns_to_vmr_ids}
\alias{all_table_ontology_columns_to_vmr_ids}
\title{Convert GRDI columns to VMR IDs.}
\usage{
all_table_ontology_columns_to_vmr_ids(db, df, vmr_table)
}
\arguments{
\item{db}{Connection to the VMR}

\item{df}{A dataframe formatted to match VMR column names}

\item{vmr_table}{The target VMR table.}
}
\description{
Convert GRDI columns that take GRDI ontologies and convert them to VMR
integer IDs.
}
