% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irida.R
\name{associate_irida_seqs_with_vmr_isolates}
\alias{associate_irida_seqs_with_vmr_isolates}
\title{Get sequence information from IRIDA for isolates in VMR without this info}
\usage{
associate_irida_seqs_with_vmr_isolates(db)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}
}
\value{
A dataframe with VMR sequence records and associated sequence
information from IRIDA
}
\description{
This function queries the VMR for isolates that have an IRIDA sample ID, and
with WGS information, but without corresponding IRIDA sequence information
(that is, irida ids representing the sequence), and attempts to return this
information from the IRIDA server.
}
