% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_for_existing_geo_loc_site}
\alias{check_for_existing_geo_loc_site}
\title{Checks for entry in geo_loc_site and adds a new one if it doesn't exist}
\usage{
check_for_existing_geo_loc_site(db, x)
}
\arguments{
\item{db}{connection to VRM}

\item{x}{Vector of GRDI column geo_loc_name (site)}
}
\description{
Checks for entry in geo_loc_site and adds a new one if it doesn't exist
}
