% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{clean_names}
\alias{clean_names}
\title{Rename the original field names to something more friendly}
\usage{
clean_names(x)
}
\arguments{
\item{x}{vector of column names}
}
\description{
Rename the original field names to something more friendly
}
