% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_reporting.R
\name{cols_all_the_same}
\alias{cols_all_the_same}
\title{Return a dataframe of columns in which all the rows are the same value}
\usage{
cols_all_the_same(df)
}
\arguments{
\item{df}{any dataframe}
}
\description{
Return a dataframe of columns in which all the rows are the same value
}
