% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiTaxon_to_vmr.R
\name{download_and_load_ncbiTaxon}
\alias{download_and_load_ncbiTaxon}
\title{Download and load the NCBITaxon ontology}
\usage{
download_and_load_ncbiTaxon()
}
\description{
Download and load the NCBITaxon ontology
}
