% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_ont_id}
\alias{extract_ont_id}
\title{Extract Ontology ID from GRDI term}
\usage{
extract_ont_id(x)
}
\arguments{
\item{x}{Vector of GRDI terms in the format "Term name [ONTOLOGY:0000000]"}
}
\description{
Extract Ontology ID from GRDI term
}
