% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{get_contact_information_id}
\alias{get_contact_information_id}
\title{Get id for contact information, inserting into the DB if necessary}
\usage{
get_contact_information_id(db, lab, name, email)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{lab}{Laboratory name, passed onto collection_information.laboratory_name}

\item{name}{Contact name, passed onto collection_information.contact_name}

\item{email}{Contact email, passed onto collection_information.contact_email}
}
\description{
Get id for contact information, inserting into the DB if necessary
}
