% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_reporting.R
\name{get_distinct_by_sample_description}
\alias{get_distinct_by_sample_description}
\title{Return distinct rows by original_sample_description}
\usage{
get_distinct_by_sample_description(df, ...)
}
\arguments{
\item{df}{Any dataframe}

\item{...}{unquoted field names to exclude from the distinct command}
}
\description{
Sorts a dataframe into distince rows based on the field original_sample_description
}
