% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_wrappers.R
\name{get_isolate_ids}
\alias{get_isolate_ids}
\title{Get isolate ids from the sample_collector_sample_ids}
\usage{
get_isolate_ids(db, isolate_ids)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{sample_names}{values to query}
}
\description{
Get isolate ids from the sample_collector_sample_ids
}
