% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mapping_for_one_col}
\alias{get_mapping_for_one_col}
\title{Get the vmr's mapping row for the GRDI column}
\usage{
get_mapping_for_one_col(db, grdi_col)
}
\description{
Get the vmr's mapping row for the GRDI column
}
