% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_sample_ids}
\alias{get_sample_ids}
\title{Get sample ids from the sample_collector_sample_ids}
\usage{
get_sample_ids(db, sample_names)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{sample_names}{values to query}
}
\description{
Get sample ids from the sample_collector_sample_ids
}
