% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_inserts.R
\name{get_unique_antibiotics}
\alias{get_unique_antibiotics}
\title{From a dataframe, extract the AMR columns and return the unique antibiotics}
\usage{
get_unique_antibiotics(df)
}
\description{
From a dataframe, extract the AMR columns and return the unique antibiotics
}
