% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{getProviderOptions}
\alias{getProviderOptions}
\title{List provider options}
\usage{
getProviderOptions(provider = "virtualbox", details = FALSE)
}
\arguments{
\item{provider}{a provider name}

\item{details}{if TRUE print options, otherwise return default options}
}
\value{
a list of options
}
\description{
List a provider available options.
}
\details{
It return a list of options name and value for a specific provider.
To get the help page do \verb{?<provider_name>Options()}, for example [\code{\link[=virtualboxOptions]{virtualboxOptions()}}].
}
\examples{
vbOpts <- getProviderOptions(provider = "virtualbox")
print(vbOpts)
}
