% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_reporting.R
\name{grdi_excel_to_sample_metadata_df}
\alias{grdi_excel_to_sample_metadata_df}
\title{Full join the Sample metadata from the Excel file template}
\usage{
grdi_excel_to_sample_metadata_df(file)
}
\arguments{
\item{file}{The path to the GRDI excel template}
}
\description{
Join the "Sample collection & Processing", "Host Information", and "Environmental conditions"
tabs from the GRDI excel template into a single dataframe, joined by sample_collector_sample_ID.
}
