% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grep_ont_term}
\alias{grep_ont_term}
\title{Search for a term in a lookup table in the VMR}
\usage{
grep_ont_term(db, lookup, x, form = c("term", "table"))
}
\arguments{
\item{db}{Connection to vmr}

\item{lookup}{lookup table to search through}

\item{x}{pattern}

\item{form}{if 'term', just return the ontology string, if table, return as a datafram}
}
\description{
Search for a term in a lookup table in the VMR
}
