% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{host_organisms_to_ids}
\alias{host_organisms_to_ids}
\title{Take the common and scientific name columns and convert to a single ID}
\usage{
host_organisms_to_ids(db, common_name, scientific_name)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{common_name}{host_common_name field}

\item{scientific_name}{host_scientific_name field}
}
\description{
Also checks to make sure that those two columns were the same ontology
term, otherwise, it is probably an error.
}
