% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_helpers.R
\name{if_len_one_rep}
\alias{if_len_one_rep}
\title{Repeat value of length 1 to max length of an argument list}
\usage{
if_len_one_rep(x)
}
\description{
If any in a list of arguments has length 1, then repeat this
value to the same length as the maxium length in the list. Required
because binding a list of parameters to an SQL query requires that
they all be the same length
}
