% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensititre_import.R
\name{import_and_insert_sensititre}
\alias{import_and_insert_sensititre}
\title{Import sensititre data, add relevant testing information, and insert to database}
\usage{
import_and_insert_sensititre(
  db,
  file,
  sheet = NA,
  testing_date = NA,
  lab_name = "NML St-Hyacinthe",
  name = "Audrey Charlebois",
  email = "audrey.charlebois@phac-aspc.gc.ca",
  agency = "Public Health Agency of Canada (PHAC) [GENEPIO:0100551]"
)
}
\arguments{
\item{db}{database connection to the vmr}

\item{file}{path to the raw file}

\item{sheet}{if the path is to an excel file, provide the sheet}

\item{testing_date}{provide a date that the samples were tested, if no column is available in the datafile}

\item{lab_name}{Name of the testing lab}

\item{name}{name of the contact who did the testing}

\item{email}{the email of the contact}

\item{agency}{the agency that did the testing.}
}
\value{
nothing, but inserts into the database without commiting it.
}
\description{
This function attempts to be a complete wrapper for the insertion of MIC
data from a sensititre raw data file into the VMR.
}
\details{
This function, as is, makes several assumptions about the breakpoints
and testing information. Breakpoints are currently an internal function
with breakpoints received from the lab of Audrey Charlebois. It will not
be suitable for data from other labs that may use different breakpoints.

Testing contact information and the date testing can be supplied manually
}
