% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_alternative_sample_ids}
\alias{insert_alternative_sample_ids}
\title{Insert alternative sample ids}
\usage{
insert_alternative_sample_ids(db, vmr_sample_id, alt_id, note = NA)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{vmr_sample_id}{vector of VMR sample_ids}

\item{alt_id}{The alternative ids of the samples}

\item{note}{A note describing the alternative ID}
}
\description{
Insert alternative sample ids
}
