% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_inserts.R
\name{insert_amr_from_wide_form_template}
\alias{insert_amr_from_wide_form_template}
\title{Insert AST data into the database based on the wide form template}
\usage{
insert_amr_from_wide_form_template(db, df)
}
\description{
Insert AST data into the database based on the wide form template
}
