% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_anatomical_data}
\alias{insert_anatomical_data}
\title{Insert data into the anatomical data table}
\usage{
insert_anatomical_data(db, sample_id, anatomical_region = NA)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}
}
\description{
Insert data into the anatomical data table
}
