% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_anatomical_multi_choices}
\alias{insert_anatomical_multi_choices}
\title{Insert into the multi-choice tables for the anatomical data}
\usage{
insert_anatomical_multi_choices(
  db,
  anatomy_ids,
  body_product = NA,
  anatomical_material = NA,
  anatomical_part = NA
)
}
\description{
Insert into the multi-choice tables for the anatomical data
}
