% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_collection_information}
\alias{insert_collection_information}
\title{Insert data into collection_information and associated tables}
\usage{
insert_collection_information(
  db,
  sample_id,
  sample_collected_by = NA,
  contact_information = NA,
  sample_collection_date = NA,
  sample_collection_date_precision = NA,
  presampling_activity_details = NA,
  sample_received_date = NA,
  original_sample_description = NA,
  specimen_processing = NA,
  sample_storage_method = NA,
  sample_storage_medium = NA,
  collection_device = NA,
  collection_method = NA
)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}
}
\description{
Inserts data into the table "collection_information",
}
