% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_env_data}
\alias{insert_env_data}
\title{Insert data into the environmental data table}
\usage{
insert_env_data(
  db,
  sample_id,
  air_temperature = NA,
  air_temperature_units = NA,
  water_temperature = NA,
  water_temperature_units = NA,
  sediment_depth = NA,
  sediment_depth_units = NA,
  water_depth = NA,
  water_depth_units = NA,
  available_data_type_details = NA
)
}
\description{
Insert data into the environmental data table
}
