% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_env_multi_choices}
\alias{insert_env_multi_choices}
\title{Insert into the multi-choice tables for the environmental data}
\usage{
insert_env_multi_choices(
  db,
  env_data_id,
  animal_or_plant_population = NA,
  available_data_types = NA,
  environmental_material_constituent = NA,
  environmental_material = NA,
  environmental_site = NA,
  weather_type = NA
)
}
\description{
Insert into the multi-choice tables for the environmental data
}
