% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_food_data}
\alias{insert_food_data}
\title{Insert data into the food data table}
\usage{
insert_food_data(
  db,
  sample_id,
  food_product_production_stream = NA,
  food_product_origin_country = NA,
  food_packaging_date = NA,
  food_quality_date = NA
)
}
\description{
Insert data into the food data table
}
