% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_geo_loc}
\alias{insert_geo_loc}
\title{Insert data into geo_loc and associated tables}
\usage{
insert_geo_loc(
  db,
  sample_id,
  country = NA,
  state_province_region = NA,
  site = NA,
  latitude = NA,
  longitude = NA
)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{country}{As ontology terms with IDS}

\item{state_province_region}{As ontology terms with IDS}

\item{site}{as VMR site ID from}

\item{latitude}{as point}

\item{lontitude}{as point}
}
\description{
Inserts data into the table "geo_loc".
}
