% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_host_data}
\alias{insert_host_data}
\title{Insert data into the hosts table}
\usage{
insert_host_data(
  db,
  sample_id,
  host_organism = NA,
  host_ecotype = NA,
  host_breed = NA,
  host_food_production_name = NA,
  host_disease = NA,
  host_age_bin = NA,
  host_origin_geo_loc_name_country = NA
)
}
\description{
Insert data into the hosts table
}
