% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_into_multi_choice_table}
\alias{insert_into_multi_choice_table}
\title{Insert df into one of the multi choice tables}
\usage{
insert_into_multi_choice_table(db, df, table)
}
\description{
Insert df into one of the multi choice tables
}
