% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_isolate_data}
\alias{insert_isolate_data}
\title{Insert new data into isolate table of the VMR}
\usage{
insert_isolate_data(
  db,
  sample_id,
  isolate_id,
  organism = NA,
  strain = NA,
  microbiological_method = NA,
  progeny_isolate_id = NA,
  isolated_by = NA,
  contact_information = NA,
  isolation_date = NA,
  isolate_received_date = NA,
  taxonomic_identification_process = NA,
  taxonomic_identification_process_details = NA,
  serovar = NA,
  serotyping_method = NA,
  phagetype = NA,
  irida_sample_id = NA,
  irida_project_id = NA,
  biosample_id = NA,
  bioproject_id = NA
)
}
\description{
Insert new data into isolate table of the VMR
}
