% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_or_return_geo_site}
\alias{insert_or_return_geo_site}
\title{Replace site values with the geo_loc_site ID, or insert and return if not present}
\usage{
insert_or_return_geo_site(db, x)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{x}{vector of string-form geo_loc_site names}
}
\description{
Replace site values with the geo_loc_site ID, or insert and return if not present
}
