% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{insert_sample_metadata}
\alias{insert_sample_metadata}
\title{Insert sample metadata}
\usage{
insert_sample_metadata(db, df)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{df}{dataframe with sample metadata columns to insert}
}
\description{
Insert sample metadata
}
