% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_wgs_ext}
\alias{insert_wgs_ext}
\title{Insert a WGS record, tying extraction to isolate}
\usage{
insert_wgs_ext(db, extraction_id, isolate_id)
}
\description{
Insert a WGS record, tying extraction to isolate
}
