% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensititre_import.R
\name{internal_breakpoints}
\alias{internal_breakpoints}
\title{A set of internal breakpoints to assign to the data}
\usage{
internal_breakpoints(x)
}
\description{
These breakpoints are based off of reports from the lab of Audrey Charlebois.
}
