% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irida.R
\name{isolates_with_no_irida_sequencing_ids}
\alias{isolates_with_no_irida_sequencing_ids}
\title{Get isolates that have an IRIDA id but are missing corresponding IRIDA sequence information}
\usage{
isolates_with_no_irida_sequencing_ids(db)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}
}
\value{
df with irida_id, sequence_id, library_id, and the sequencing_platform
}
\description{
Get isolates that have an IRIDA id but are missing corresponding IRIDA sequence information
}
