% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_inserts.R
\name{long_form_antibiotics}
\alias{long_form_antibiotics}
\title{From a wide-form grdi spec df, return the AMR fields as a long form df}
\usage{
long_form_antibiotics(df)
}
\description{
From a wide-form grdi spec df, return the AMR fields as a long form df
}
