% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lookup_table_terms}
\alias{lookup_table_terms}
\title{Get a table of ontology terms from a lookup table}
\usage{
lookup_table_terms(db, lookup_table)
}
\description{
Get a table of ontology terms from a lookup table
}
