% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_schema.R
\name{make_blank_df}
\alias{make_blank_df}
\title{Make a blank dataframe with the template column names}
\usage{
make_blank_df(nrow = 1, type = c("samples", "isolates", "sequences"))
}
\arguments{
\item{nrow}{number of rows for the dataframe}
}
\value{
empty dataframe of length nrow
}
\description{
All will be NA
}
