% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_helpers.R
\name{make_sql_params}
\alias{make_sql_params}
\title{Return an SQL formatted list of args to pass to an insert statement}
\usage{
make_sql_params(n)
}
\description{
Return an SQL formatted list of args to pass to an insert statement
}
