% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{multi_choice_tables}
\alias{multi_choice_tables}
\title{Get a list of multi choice tables associated with the VMR table}
\usage{
multi_choice_tables(db, table)
}
\description{
Get a list of multi choice tables associated with the VMR table
}
