% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{new_extraction}
\alias{new_extraction}
\title{Insert an extraction record}
\usage{
new_extraction(
  db,
  experimental_protocol_field = NA,
  experimental_specimen_role_type = NA,
  nucleic_acid_extraction_method = NA,
  nucleic_acid_extraction_kit = NA,
  sample_volume_measurement_value = NA,
  sample_volume_measurement_unit = NA,
  residual_sample_status = NA,
  sample_storage_duration_value = NA,
  sample_storage_duration_unit = NA,
  nucleic_acid_storage_duration_value = NA,
  nucleic_acid_storage_duration_unit = NA
)
}
\description{
Insert an extraction record
}
