% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{new_project}
\alias{new_project}
\title{Insert new project into the VMR database}
\usage{
new_project(
  db,
  sample_plan_id,
  sample_plan_name,
  project_name = NA,
  description = NA
)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{sample_plan_id}{VMR sample plan id}

\item{sample_plan_name}{The sample plan name}

\item{project_name}{The project name}

\item{description}{Description of the project}
}
\description{
Insert new project into the VMR database
}
