% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{new_sequence}
\alias{new_sequence}
\title{Insert a sequence record}
\usage{
new_sequence(
  db,
  extraction_id = NA,
  library_id = NA,
  contact_information = NA,
  sequenced_by = NA,
  sequencing_date = NA,
  sequencing_project_name = NA,
  sequencing_platform = NA,
  sequencing_instrument = NA,
  sequencing_assay_type = NA,
  dna_fragment_length = NA,
  genomic_target_enrichment_method = NA,
  genomic_target_enrichment_method_details = NA,
  amplicon_pcr_primer_scheme = NA,
  amplicon_size = NA,
  sequencing_flow_cell_version = NA,
  library_preparation_kit = NA,
  sequencing_protocol = NA,
  r1_fastq_filename = NA,
  r2_fastq_filename = NA,
  fast5_filename = NA,
  r1_irida_id = NA,
  r2_irida_id = NA
)
}
\description{
Insert a sequence record
}
