% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{open_sheet}
\alias{open_sheet}
\title{Open a sheet from the Excel Template and return a neat dataframe}
\usage{
open_sheet(file, sheet_name)
}
\arguments{
\item{file}{Path to the excel template file with data}

\item{sheet_name}{The excel sheet to read from.}
}
\description{
Open a sheet from the Excel Template and return a neat dataframe
}
