% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_schema.R
\name{parse_latest_grdi_schema}
\alias{parse_latest_grdi_schema}
\title{Get the latest GRDI schema as an R function}
\usage{
parse_latest_grdi_schema()
}
\description{
Get the latest GRDI schema as an R function
}
