% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{print.vmr}
\alias{print.vmr}
\title{Print \strong{vmr} object information}
\usage{
\method{print}{vmr}(x, ...)
}
\arguments{
\item{x}{a \strong{vmr} object}

\item{...}{optional print arguments}
}
\value{
the \strong{vmr} object (via invisible(x))
}
\description{
print information from a \strong{vmr} object
}
