% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print_tallies_of_columns}
\alias{print_tallies_of_columns}
\title{Print out a tally of unique values across all columns in a dataframe}
\usage{
print_tallies_of_columns(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Except for ID columns!
}
