% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irida.R
\name{query_pairs_from_irida}
\alias{query_pairs_from_irida}
\title{Get paired-end reads from IRIDA}
\usage{
query_pairs_from_irida(samples)
}
\arguments{
\item{samples}{a vector of irida sample ID.}
}
\value{
dataframe of paired-end reads assocaited with each sample
}
\description{
Uses \link[RiRida:get_sequences]{RiRida::get_sequences} to query IRIDA for all paired-end reads for the
supplied samples, and returns a formatted DF.
}
