% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{read_template_from_excel}
\alias{read_template_from_excel}
\title{Read fields from a GRDI excel template, combining sheets if necessary}
\usage{
read_template_from_excel(file, fields = c("Samples", "Isolates", "Sequences"))
}
\arguments{
\item{file}{excel file to read}

\item{fields}{One of "Samples", "Isolates", or "Sequences"}
}
\description{
Read fields from a GRDI excel template, combining sheets if necessary
}
