% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{rename_grdi_cols_to_vmr}
\alias{rename_grdi_cols_to_vmr}
\title{Renaming a GRDI formatted dataframe to a VMR table}
\usage{
rename_grdi_cols_to_vmr(db, df, table)
}
\arguments{
\item{db}{A DBI connection to the VMR}

\item{df}{Dataframe, formatted according to the GRDI spec}

\item{table}{The table in the VMR database to convert into}
}
\value{
A dataframe with columns that match the fields of the selected
table in the VMR
}
\description{
Renaming a GRDI formatted dataframe to a VMR table
}
