% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_and_inserts.R
\name{repIns_collection_info}
\alias{repIns_collection_info}
\title{Report and insert collection info}
\usage{
repIns_collection_info(db, sample_df, contact_id)
}
\description{
Report and insert collection info
}
