% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{return_or_insert_contact_information}
\alias{return_or_insert_contact_information}
\title{Get id for contact information, inserting into the DB if necessary}
\usage{
return_or_insert_contact_information(
  db,
  lab = "Not Provided [GENEPIO:0001668]",
  name = "Not Provided [GENEPIO:0001668]",
  email = "Not Provided [GENEPIO:0001668]",
  note = NA
)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{lab}{Laboratory name, passed onto collection_information.laboratory_name}

\item{name}{Contact name, passed onto collection_information.contact_name}

\item{email}{Contact email, passed onto collection_information.contact_email}
}
\description{
Get id for contact information, inserting into the DB if necessary
}
