% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_reporting.R
\name{save_report_to_excel_workbook}
\alias{save_report_to_excel_workbook}
\title{Format an Excel GRDI file to report on same-value columns and distinct metadata per sample description}
\usage{
save_report_to_excel_workbook(file, title, same_df, grouped_df)
}
\arguments{
\item{file}{Output file}

\item{title}{Title of the excel workbook}

\item{same_df}{The same_df dataframe returned from \link{samples_by_group}}

\item{grouped_df}{The grouped_df dataframe returned from \link{samples_by_group}}
}
\description{
Format an Excel GRDI file to report on same-value columns and distinct metadata per sample description
}
