% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amr_inserts.R
\name{select_by_antibiotic}
\alias{select_by_antibiotic}
\title{Select the amr fields based of a single antibiotic}
\usage{
select_by_antibiotic(df, antibiotic)
}
\description{
Select the amr fields based of a single antibiotic
}
