% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_fields}
\alias{select_fields}
\title{From a DF, select only those cols relevant to the table insert.}
\usage{
select_fields(db, db_table, df_cols)
}
\description{
Also: warn if there are columns in the database that are missing in the
input dataframe, but do not throw an error.
}
