% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{seperate_column_into_longform}
\alias{seperate_column_into_longform}
\title{Separate a GRDI multi choice column into a long-form dataframe}
\usage{
seperate_column_into_longform(df, col)
}
\description{
Separate a GRDI multi choice column into a long-form dataframe
}
