% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{set_vmr_isolate_id_from_alternates}
\alias{set_vmr_isolate_id_from_alternates}
\title{First, just set the respective isolate IDs to VMR isolate IDs}
\usage{
set_vmr_isolate_id_from_alternates(db, x)
}
\arguments{
\item{db}{\link{DBI} connection}

\item{x}{The vector to attempt to convert to VMR IDs}
}
\description{
Attempt to use the possible_isolate_names view to set the user
isolate IDs to VMR isolate_IDs
}
