% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_reporting.R
\name{sort_by_categories}
\alias{sort_by_categories}
\title{Sort a vector of column names into categories}
\usage{
sort_by_categories(x)
}
\arguments{
\item{x}{A vector of GRDI column names}
}
\description{
This is based on regular expressions
}
