% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_helpers.R
\name{sql_args_to_ontology_ids}
\alias{sql_args_to_ontology_ids}
\title{Convert the vectors of ontology terms into their corresponding VMR ids}
\usage{
sql_args_to_ontology_ids(db, ontology_columns, sql_arguments)
}
\arguments{
\item{ontology_columns}{A vector of the list names that are to be
converted into VMR ids}

\item{sql_arguments}{A names list of vectors, which contain the vectors
of ontology IDs to be converted into VMR ids.}
}
\value{
A named list of vector of same length as \code{sql_arguments}
}
\description{
Given a list of vectors that are to be input into the VMR, convert the
desired columns containing ontology terms into their respective VMR IDs
}
