% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_helpers.R
\name{sql_args_to_uniform_list}
\alias{sql_args_to_uniform_list}
\title{Convert a function's arguments into a list of vectors of same length}
\usage{
sql_args_to_uniform_list(env)
}
\arguments{
\item{env}{the \link{environment} call of a function}
}
\value{
a list of the functions arguments, minus the DBI connection argument
}
\description{
Input the \code{environment()} of a function to return the parameters as a
list of equal length vectors, without the db argument that is common to
all the insert functions.
}
