% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{summary.vmr}
\alias{summary.vmr}
\title{Summary \strong{vmr} object information}
\usage{
\method{summary}{vmr}(object, ...)
}
\arguments{
\item{object}{a \strong{vmr} object}

\item{...}{optional print arguments}
}
\value{
the \strong{vmr} object (via invisible(x))
}
\description{
print information from a \strong{vmr} object
}
