% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{tally_values_to_string}
\alias{tally_values_to_string}
\title{Quickly get all unique values in a vector and print them out in a string}
\usage{
tally_values_to_string(x)
}
\arguments{
\item{x}{a vector of data}
}
\description{
This is for the markdown insertion report, for inline tallies of what is
in a field of data
}
