% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_and_insert_multi_choice_column}
\alias{transform_and_insert_multi_choice_column}
\title{Populate a "multi-choice" table in the VMR}
\usage{
transform_and_insert_multi_choice_column(db, df, column)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{df}{The working dataframe that at least contains the GRDI column, as
well as the id field that the multi-choice table will use as a foreign key.}

\item{column}{str, the name of the GRDI column that can take on multiple
values

@export}
}
\description{
Currently, the GRDI schema allows some fields to be populated with multiple
values. These fields are implemented in the database as their own tables.
These tables are all similar in their structure, so this convenience
function quickly populates these tables given the values in the GRDI fields.
The function assumes that the values are separated wither with
"|", or a ";"
}
