% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irida.R
\name{update_sequences_with_irida}
\alias{update_sequences_with_irida}
\title{Update VMR with irida sequence information}
\usage{
update_sequences_with_irida(db, seq_id, f_file, r_file, f_id, r_id)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}
}
\description{
Update VMR with irida sequence information
}
