% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr.R
\docType{package}
\encoding{UTF-8}
\name{vmr-package}
\alias{vmr-package}
\alias{_PACKAGE}
\alias{vmr}
\title{Virtual Machines for R}
\description{
Manage, provision and use Virtual Machines pre-configured for R.
Develop, test and build package in a clean environment.
'Vagrant' tool and a provider (such as 'Virtualbox') have to be installed.
}
\details{
\tabular{ll}{
Package: \tab vmr\cr
Type: \tab Package\cr
Version: \tab 0.0.2\cr
Date: \tab 2021-02-19\cr
License: \tab GPL (>=3)\cr
}

This package is a wrap of the \href{https://www.vagrantup.com/}{Vagrant}
tool and more.
It allows to manage, provision and use Virtual Machines pre-configured for R.
It currently only uses 'Virtualbox' (>= 6.1.14) as provider.
Vagrant tool have to be installed too.
Used VMs come from \url{https://app.vagrantup.com/VMR}
repository and the sources use to generate them can be found at
\url{https://gitlab.com/rstuff/vms}.
See vignettes for the documentations \code{browseVignette("vmr")}.
}
\examples{
\dontrun{
library("vmr")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/rstuff/vmr}
  \item \url{https://rstuff.gitlab.io/vmr}
  \item Report bugs at \url{https://gitlab.com/rstuff/vmr/-/issues}
}

}
\author{
Jean-François Rey \email{jf.rey.public@gmail.com}

Maintainer: Jean-François Rey \email{jf.rey.public@gmail.com}
}
\keyword{machine}
\keyword{provider}
\keyword{provision}
\keyword{vagrant}
\keyword{virtual}
\keyword{virtualbox}
