% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrBoxDownload}
\alias{vmrBoxDownload}
\title{Download a Box}
\usage{
vmrBoxDownload(vmr)
}
\arguments{
\item{vmr}{a \strong{vmr} object}
}
\value{
a \strong{vmr} object
}
\description{
Download a box from a \strong{vmr} object.
}
