% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrConnect}
\alias{vmrConnect}
\title{Open a ssh connection to guest machine}
\usage{
vmrConnect(vmr)
}
\arguments{
\item{vmr}{a \strong{vmr} object}
}
\value{
a \strong{vmr} object
}
\description{
Open a ssh connection to guest machine
}
\details{
To open a ssh connection 'ssh' package have to be installed.
}
