% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrDestroy}
\alias{vmrDestroy}
\title{Remove all resources created in a \strong{vmr} environment}
\usage{
vmrDestroy(id = "", force = FALSE)
}
\arguments{
\item{id}{a \strong{vmr} environment id (default : "" id from the current environment)}

\item{force}{if TRUE force to remove}
}
\value{
the vagrant environment id
}
\description{
Remove all resources created by \code{\link[=vmrStart]{vmrStart()}}
}
\details{
Will by default remove all resources created from the current \strong{vmr} environment.
By specifying the \emph{id} any environment with this \emph{id} will be remove.
}
\examples{
\dontrun{
vmrStop()
vmrDestroy()
}
}
