% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrDisconnect}
\alias{vmrDisconnect}
\title{Disconnect ssh connection to guest machine}
\usage{
vmrDisconnect(vmr)
}
\arguments{
\item{vmr}{a \strong{vmr} object}
}
\value{
a \strong{vmr} object
}
\description{
Close a ssh connection to the guest machine
}
\details{
'ssh' package need to be installed.
}
