% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrExec}
\alias{vmrExec}
\title{Execute R methods into guest machine}
\usage{
vmrExec(cmd = c())
}
\arguments{
\item{cmd}{list of R command}
}
\value{
\code{NULL}
}
\description{
Run R method into guest machine.
}
\details{
call Rscript -e "cmd" into the guest machine from
the current \strong{vmr} environment.
Command are independents and do not keep memory of past commands.
}
\examples{
\dontrun{
cmd <- c("Sys.info()", 'print("Hello World!")')
vmrExec(cmd)
}
}
