% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrInfo}
\alias{vmrInfo}
\title{Get guest machine information}
\usage{
vmrInfo()
}
\value{
\code{NULL}
}
\description{
Get guest machine information.
Print OS, R, R-devel and R packages information.
Still in development.
}
\examples{
\dontrun{
boxes <- vmrList()
vmr <- vmrCreate(boxes$Name[1])
vmr <- vmrInitEnv(vmr)
vmrStart()
vmrInfo()
}
}
