% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrInstallPackages}
\alias{vmrInstallPackages}
\title{Install R packages into guest machine}
\usage{
vmrInstallPackages(pkgs = c())
}
\arguments{
\item{pkgs}{list of R packages}
}
\value{
installed packages vector
}
\description{
Install a list of R packages into the guest machine
of the current \strong{vmr} environment.
}
\examples{
\dontrun{
vmrInstallPackages(c("vmr"))
}
}
