% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrIsRunning}
\alias{vmrIsRunning}
\title{Is \strong{vmr} environment running}
\usage{
vmrIsRunning()
}
\value{
TRUE if running, otherwise FALSE
}
\description{
Check if a guest machine in a \strong{vmr} environment is running
}
\examples{
\dontrun{
lboxes <- vmrList()
vmr <- vmrCreate(lboxes$Name[1])
vmr <- vmrInitEnv(vmr)
vmrStart()
vmrIsRunning()
vmrStop()
vmrIsRunning()
}
}
