% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrList}
\alias{vmrList}
\title{List available boxes from VagrantCloud}
\usage{
vmrList(org = .VagrantCloudOrganization)
}
\arguments{
\item{org}{Vagrant Cloud organization name (default : 'VMR')}
}
\value{
a data.frame with Name, Provider, Version and Description of available boxes
}
\description{
List of available boxes from a VagrantCloud organization account.
}
\details{
Default usage lists boxes preconfigurated with R
from \href{https://app.vagrantup.com/VMR}{VMR organization account}.
}
