% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrListBox}
\alias{vmrListBox}
\title{List all available version of a box}
\usage{
vmrListBox(box_name, org = .VagrantCloudOrganization)
}
\arguments{
\item{box_name}{the box name}

\item{org}{Vagrant Cloud organization name (default : 'VMR')}
}
\value{
a data.frame with "Name, "Version", "Description", "Provider" and "Date" of the box
}
\description{
List all versions and providers available of a box.
}
\details{
List information of a box from VagrantCloud.
Default usage list information of a box preconfigurated with R
from \href{https://app.vagrantup.com/VMR}{VMR organization account}.
}
\examples{
\dontrun{
# List Boxes
boxes <- vmrList()
# Box informaion
box_info <- vmrListBox(boxes$Name[1])
box_info
}
}
