% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrLoad}
\alias{vmrLoad}
\title{Load a \strong{vmr} environment containing a Vagrant file}
\usage{
vmrLoad(dir = "./", vagrantfileName = "Vagrantfile")
}
\arguments{
\item{dir}{the \strong{vmr} environment directory (default: "./")}

\item{vagrantfileName}{a Vagrantfile name (default: "Vagrantfile")}
}
\value{
a \strong{vmr} object
}
\description{
Load a \strong{vmr} environment containing a VagrantFile
and create a \strong{vmr} object (see [\code{\link[=vmrCreate]{vmrCreate()}}] for object details).
}
\details{
It read a Vagrant file template with \strong{vmr} compatible parameters.
It's an experimental Vagrant file reading, some parameters may not be loaded.
}
\examples{
\dontrun{
# load the Vagrantfile in the current directory
vmr <- vmrLoad(getwd())
}
}
