% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrLocalBoxList}
\alias{vmrLocalBoxList}
\title{List downloaded boxes}
\usage{
vmrLocalBoxList()
}
\value{
a data.frame with boxes Name, Providers and Version
}
\description{
List all boxes downloaded in localhost
}
\examples{
\dontrun{
localBoxes <- vmrLocalBoxList()
print(localBoxes)
}
}
