% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrLocalBoxPrune}
\alias{vmrLocalBoxPrune}
\title{Remove old installed boxes}
\usage{
vmrLocalBoxPrune()
}
\value{
a data.frame of still installed boxes (Name, Poviders and Version)
}
\description{
Removes old versions of installed boxes.
}
\examples{
\dontrun{
vmrLocalBoxPrune()
}
}
