% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrLocalBoxRemove}
\alias{vmrLocalBoxRemove}
\title{Remove a box from localhost}
\usage{
vmrLocalBoxRemove(name, provider = "", version = "", force = FALSE)
}
\arguments{
\item{name}{the box name}

\item{provider}{the box provider (default: first provider found)}

\item{version}{the box version (default: version available)}

\item{force}{if TRUE force to remove}
}
\value{
execution code or message
}
\description{
Remove a specific box from localhost.
}
\examples{
\dontrun{
lboxes <- vmrLocalBoxList()
vmrLocalBoxRemove(lboxes$Name[[1]])
# if multiple providers and versions
vmrLocalBoxRemove(lboxes$Name[[1]], lboxes$Provider[[1]], lboxes$Version[[1]])
}
}
