% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrLocalBoxUpdate}
\alias{vmrLocalBoxUpdate}
\title{Update local box version}
\usage{
vmrLocalBoxUpdate()
}
\value{
execution code or message
}
\description{
Download the latest version of the box use in the current \strong{vmr} environment.
}
