% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrPackageBuild}
\alias{vmrPackageBuild}
\title{Build a package in the guest machine}
\usage{
vmrPackageBuild(pkg = "./", binary = FALSE)
}
\arguments{
\item{pkg}{a package directory or a tar.gz file}

\item{binary}{if TRUE build binary package otherwise FALSE}
}
\value{
\code{NULL}
}
\description{
Build a package bundle or binary into the guest machine.
}
\details{
upload the package and run \code{\link[devtools:build]{devtools::build()}}
(build available in $HOME/vmr/package/pkg) in the current \strong{vmr} environment.
}
