% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrPackageCheck}
\alias{vmrPackageCheck}
\title{Perform a package check on guest}
\usage{
vmrPackageCheck(pkg = "./")
}
\arguments{
\item{pkg}{a package directory or a tar.gz file}
}
\value{
\code{NULL}
}
\description{
Perform a package check into the guest
}
\details{
upload the package and run devtools::check()
into the guest machine. (check available in $HOME/vmr/package/pkg).
Checking a directory with multiple files may slower upload, prefer tar.gz file
}
\examples{
\dontrun{
vmrPackageCheck("vmr_package.tar.gz")
}
}
