% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrPackageTest}
\alias{vmrPackageTest}
\title{Test a package into a guest machine}
\usage{
vmrPackageTest(pkg = "./")
}
\arguments{
\item{pkg}{a package directory or tar.gz}
}
\value{
\code{NULL}
}
\description{
Test a package into a guest machine
}
\details{
Perform a package check into the guest machine
of the current \strong{vmr} environment using \code{\link[devtools:test]{devtools::test()}}.
(tests are available in $HOME/vmr/package/pkg)
}
