% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrProvision}
\alias{vmrProvision}
\title{Provision a \strong{vmr} environment}
\usage{
vmrProvision(cmd = c(), elts = c(), dest = "")
}
\arguments{
\item{cmd}{list of shell commands}

\item{elts}{list of files and/or directories}

\item{dest}{destination of elts (default HOME/vmr)}
}
\value{
\code{NULL}
}
\description{
Provision a \strong{vmr} environment.
}
\details{
Upload 'elts' files and/or directories to the guest machine 'dest'
from the current \strong{vmr} environment.
And finaly run shell commands 'cmd' in the guest machine.
}
