% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrRemoveSnapshot}
\alias{vmrRemoveSnapshot}
\title{remove a snapshot of the guest machine}
\usage{
vmrRemoveSnapshot(snap_name)
}
\arguments{
\item{snap_name}{the snapshot name}
}
\value{
\code{NULL}
}
\description{
remove a snapshot of the guest machine
}
