% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrRestoreSnapshot}
\alias{vmrRestoreSnapshot}
\title{Restore a snapshot of the guest machine}
\usage{
vmrRestoreSnapshot(snap_name)
}
\arguments{
\item{snap_name}{the snapshot name}
}
\value{
the snapshot name
}
\description{
Restore a snapshot of the guest machine.
}
\examples{
\dontrun{
vmrRestoreSnapshot("my snapshot")
}
}
