% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrResume}
\alias{vmrResume}
\title{Resume a stopped guest machine}
\usage{
vmrResume()
}
\value{
\code{NULL}
}
\description{
Resume a stopped guest machine.
}
\details{
In the current \strong{vmr} environment, start a stopped ([\code{\link[=vmrSuspend]{vmrSuspend()}}]) guest machine.
}
