% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrSend}
\alias{vmrSend}
\title{Send files and/or directories to guest machine}
\usage{
vmrSend(elt = c())
}
\arguments{
\item{elt}{list of files and directories}
}
\value{
0 if OK, message otherwise
}
\description{
Send files and/or directories to the guest machine
in the current \strong{vmr} environment.
They are upload into ~/vmr/ directory.
}
\examples{
\dontrun{
vmrSend(c("myfile"))
}
}
