% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrSetVerbose}
\alias{vmrSetVerbose}
\title{Set verbose level}
\usage{
vmrSetVerbose(verbose_mode = "Normal")
}
\arguments{
\item{verbose_mode}{"None", "Normal" or "Full"}
}
\value{
invisible verbose value
}
\description{
Set verbose level for vmr package functions
}
\details{
Three verboses mode is available:
\itemize{
\item "None" : print nothings
\item "Normal" : print essential
\item "Full" : print all
}
}
