% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrStart}
\alias{vmrStart}
\title{Start a \strong{vmr} environment}
\usage{
vmrStart()
}
\value{
the vmr environment unique id
}
\description{
Start a guest virtual machine using the current \strong{vmr} environment
(directory and Vagrantfile template)
}
\examples{
\dontrun{
lboxes <- vmrList()
vmr <- vmrCreate(lboxes$Name[1])
vmr <- vmrInitEnv(vmr)
vmrStart()
vmrStop()
}
}
