% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrStatus}
\alias{vmrStatus}
\title{Get the state of the guest machine}
\usage{
vmrStatus()
}
\value{
a data.frame with Name, Provider and state
}
\description{
Print guest machine state in the current \strong{vmr} environment.
}
