% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrStop}
\alias{vmrStop}
\title{Stop a \strong{vmr} environement}
\usage{
vmrStop(force = FALSE)
}
\arguments{
\item{force}{if TRUE force to stop (powerOff), otherwise FALSE clean shutdown}
}
\value{
\code{NULL}
}
\description{
Stop a guest virtual machine in the current \strong{vmr} environment.
}
\examples{
\dontrun{
lboxes <- vmrList()
vmr <- vmrCreate(lboxes$Name[1])
vmr <- vmrInitEnv(vmr)
vmrStart()
vmrStop()
}
}
