% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrSuspend}
\alias{vmrSuspend}
\title{Save state and stop guest machine}
\usage{
vmrSuspend()
}
\value{
\code{NULL}
}
\description{
Save the guest machine and stop it.
}
\details{
In the current \strong{vmr} environment, save the state of the guest machine
and stop it.
}
