% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrTakeSnapshot}
\alias{vmrTakeSnapshot}
\title{Take a snapshot of the guest machine}
\usage{
vmrTakeSnapshot(snap_name)
}
\arguments{
\item{snap_name}{the name given to the snapshot}
}
\value{
the snapshot name (invisible)
}
\description{
Take a snapshot of the guest machine.
}
\examples{
\dontrun{
vmrTakeSnapshot("my snapshot")
}
}
