% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmr-methods.R
\name{vmrUpdatePackages}
\alias{vmrUpdatePackages}
\title{Update R packages installed}
\usage{
vmrUpdatePackages()
}
\description{
Updates R packages installed in the guest machine.
}
\details{
Will perform a \code{\link[=update.packages]{update.packages()}} in the guest machine
of the current \strong{vmr} environment.
}
\examples{
\dontrun{
lboxes <- vmrList()
vmr <- vmrCreate(lboxes$Name[1])
vmr <- vmrInitEnv(vmr)
vmrStart()
vmrUpdatePackages()
}
}
