% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbiTaxon_to_vmr.R
\name{append_ncbi_taxon_ontology_to_microbes}
\alias{append_ncbi_taxon_ontology_to_microbes}
\title{Get full NCBITaxon ontology and add bacterial genera and species to microbes table}
\usage{
append_ncbi_taxon_ontology_to_microbes(db, ncbi, insert = FALSE)
}
\arguments{
\item{db}{DBI connection to VMR}

\item{ncbi}{ontology loaded into R}

\item{commit}{Commit to the VMR?}
}
\description{
Get full NCBITaxon ontology and add bacterial genera and species to microbes table
}
