% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_alternative_isolate_ids}
\alias{insert_alternative_isolate_ids}
\title{Convenience function to update alternative isolate IDS with notes}
\usage{
insert_alternative_isolate_ids(db, sample_ids, iso_ids, alt_ids, notes)
}
\description{
Convenience function to update alternative isolate IDS with notes
}
