% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_collection_information}
\alias{insert_collection_information}
\title{Insert data into collection_information and associated tables}
\usage{
insert_collection_information(db, df)
}
\arguments{
\item{db}{a \link{DBI} connection to the VMR}

\item{df}{a dataframe of the relevant GRDI fields.}
}
\description{
Inserts data into the table "collection_information", and the
associated multi-choice tables "sample_purpose" and "sample_activity".
}
