% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserts.R
\name{insert_geo_loc}
\alias{insert_geo_loc}
\title{Insert data into geo_loc and associated tables}
\usage{
insert_geo_loc(db, df)
}
\arguments{
\item{db}{a \link{DBI} connection to the VMR}

\item{df}{a dataframe of the relevant GRDI fields.}
}
\description{
Inserts data into the table "geo_loc". Will check for existing entry
}
