% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_into_multi_choice_table}
\alias{insert_into_multi_choice_table}
\title{Insert values into one of the common-format multi-choice metadata tables}
\usage{
insert_into_multi_choice_table(db, ids, vals, table, is_ontology = FALSE)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{ids}{a vector of the foreign-key ids of the table that the multi-choice
table is related by.}

\item{vals}{a vector of values, of same length as ids, that are to be inserted
as values into the multi-choice table}

\item{table}{The name of the destination multi-choice table}

\item{is_ontology}{Are the values GRDI ontology terms?}
}
\description{
Currently, the GRDI schema allows some fields to be populated with multiple
values. These fields are implemented in the database as their own tables.
These tables are all similar in their structure, so this convenience
function quickly populates these tables given the values in the GRDI fields.
The function assumes that the values are separated wither with
"|", or a ";"
}
