% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_helpers.R
\name{make_insert_sql}
\alias{make_insert_sql}
\title{Create a generic SQL insert statement}
\usage{
make_insert_sql(table_name, field_names)
}
\description{
Makes a generate SQL insert statement from a table name and a list
of parameters, which should correspond to that table's fields.
}
