% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_sql_params}
\alias{make_sql_params}
\title{Return an SQL formatted list of postgres placeholders to pass to an insert statement}
\usage{
make_sql_params(n)
}
\arguments{
\item{n}{number of placeholders}
}
\description{
Return an SQL formatted list of postgres placeholders to pass to an insert statement
}
