% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{new_project}
\alias{new_project}
\title{Insert new project into the VMR database}
\usage{
new_project(db, df, description = NA)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{df}{the dataframe with the project columns sample plan id}

\item{description}{Description of the project}
}
\description{
Insert new project into the VMR database
}
