% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_reporting.R
\name{samples_by_group}
\alias{samples_by_group}
\title{Return distinct rows by original_sample_description AND fields with a single value}
\usage{
samples_by_group(df)
}
\arguments{
\item{df}{Dataframe of the GRDI template with all the sample columns}
}
\description{
Return distinct rows by original_sample_description AND fields with a single value
}
