% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irida.R
\name{update_sequences_with_irida}
\alias{update_sequences_with_irida}
\title{Update VMR with irida sequence information}
\usage{
update_sequences_with_irida(db, df)
}
\arguments{
\item{db}{\link{DBI} connection to VMR}

\item{df}{dataframe returned from \link{associate_irida_seqs_with_vmr_isolates}}
}
\description{
Update VMR with irida sequence information
}
